backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-multipart
openai
langchain
requests
python-dotenv
google-generativeai
